Long term LFP

Closes #7

See merge request sklemm/florawebplus!1

+++ app/Http/Controllers/Admin/ColumnController.php
@@ -7,8 +7,10 @@ use App\Selectlist;
 use App\Element;
 use App\Attribute;
 use App\Value;
+use App\Utils\Localization;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 use Redirect;
 
 class ColumnController extends Controller
@@ -45,30 +47,22 @@ class ColumnController extends Controller
         $lists = Selectlist::where('internal', false)->orderBy('name')->get();
         
         // Get current UI language
-        $lang = 'name_'. app()->getLocale();
+        $lang = app()->getLocale();
         
         // Get name attribute for current language
-        $attribute = Attribute::where('name', $lang)->first();
+        $attribute = Attribute::where('name', 'name_'. $lang)->first();
+        
+        // Get localized names of columns
+        $translations = Localization::getTranslations($lang, 'name');
         
         // Get data types of columns with localized names
-        $data_types = Value::whereHas('element', function ($query) {
-            $query->where('list_fk', Selectlist::where('name', '_data_type_')->first()->list_id);
-        })
-        ->whereHas('attribute', function ($query) use ($lang) {
-            $query->where('name', $lang);
-        })
-        ->orderBy('value')->get();
+        $data_types = Localization::getDataTypes($lang);
         
-        // Get with localized names of columns
-        $translations = Value::whereHas('element', function ($query) {
-            $query->where('list_fk', Selectlist::where('name', '_translation_')->first()->list_id);
-        })
-        ->whereHas('attribute', function ($query) use ($lang) {
-            $query->where('name', $lang);
-        })
-        ->orderBy('value')->get();
+        // Get IDs for data_types
+        $data_type_ids['_list_'] = Value::where('value', '_list_')->first()->element_fk;
+        $data_type_ids['_multi_list_'] = Value::where('value', '_multi_list_')->first()->element_fk;
         
-        return view('admin.column.create', compact('lists', 'data_types', 'translations', 'attribute'));
+        return view('admin.column.create', compact('lists', 'data_types', 'data_type_ids', 'translations', 'attribute'));
     }
 
     /**
@@ -79,12 +73,19 @@ class ColumnController extends Controller
      */
     public function store(Request $request)
     {
-        // Get the element_id of data_type '_list' for validating the user's input
-        // TODO: check for list '_data_type' and maybe attribute
-        $value = Value::where('value', '_list_')->first();
-        
         $request->validate([
-            'list' => 'required_if:data_type,'.$value->element_fk,
+            'list' => [
+                Rule::requiredIf(function () use ($request) {
+                    if ($request->input('data_type') == Value::where('value', '_list_')->first()->element_fk) {
+                        return true;
+                    }
+                    if ($request->input('data_type') == Value::where('value', '_multi_list_')->first()->element_fk) {
+                        return true;
+                    }
+                    return false;
+                }),
+                'integer',
+            ],
             'data_type' => 'required|integer',
             'translation' => 'required|integer',
             'description' => 'required|string',
@@ -93,7 +94,7 @@ class ColumnController extends Controller
         ]);
         
         $data = [
-            'list_fk' => ($request->input('data_type') == $value->element_fk) ? $request->input('list') : null,
+            'list_fk' => $this->getListIdFromFormRequest($request),
             'data_type_fk' => $request->input('data_type'),
             'translation_fk' => $request->input('translation'),
             'description' => $request->input('description'),
@@ -146,27 +147,19 @@ class ColumnController extends Controller
         $lists = Selectlist::where('internal', false)->orderBy('name')->get();
         
         // Get current UI language
-        $lang = 'name_'. app()->getLocale();
+        $lang = app()->getLocale();
+        
+        // Get localized names of columns
+        $translations = Localization::getTranslations($lang, 'name');
         
         // Get data types of columns with localized names
-        $data_types = Value::whereHas('element', function ($query) {
-            $query->where('list_fk', Selectlist::where('name', '_data_type_')->first()->list_id);
-        })
-        ->whereHas('attribute', function ($query) use ($lang) {
-            $query->where('name', $lang);
-        })
-        ->orderBy('value')->get();
+        $data_types = Localization::getDataTypes($lang);
         
-        // Get with localized names of columns
-        $translations = Value::whereHas('element', function ($query) {
-            $query->where('list_fk', Selectlist::where('name', '_translation_')->first()->list_id);
-        })
-        ->whereHas('attribute', function ($query) use ($lang) {
-            $query->where('name', $lang);
-        })
-        ->orderBy('value')->get();
+        // Get IDs for data_types
+        $data_type_ids['_list_'] = Value::where('value', '_list_')->first()->element_fk;
+        $data_type_ids['_multi_list_'] = Value::where('value', '_multi_list_')->first()->element_fk;
         
-        return view('admin.column.edit', compact('column', 'lists', 'data_types', 'translations'));
+        return view('admin.column.edit', compact('column', 'lists', 'data_types', 'data_type_ids', 'translations'));
     }
 
     /**
@@ -178,18 +171,25 @@ class ColumnController extends Controller
      */
     public function update(Request $request, Column $column)
     {
-        // Get the element_id of data_type '_list' for validating the user's input
-        // TODO: check for list '_data_type' and maybe attribute
-        $value = Value::where('value', '_list_')->first();
-        
         $request->validate([
-            'list' => 'required_if:data_type,'.$value->element_fk,
+            'list' => [
+                Rule::requiredIf(function () use ($request) {
+                    if ($request->input('data_type') == Value::where('value', '_list_')->first()->element_fk) {
+                        return true;
+                    }
+                    if ($request->input('data_type') == Value::where('value', '_multi_list_')->first()->element_fk) {
+                        return true;
+                    }
+                    return false;
+                }),
+                'integer',
+            ],
             'data_type' => 'required|integer',
             'translation' => 'required|integer',
             'description' => 'required|string',
         ]);
         
-        $column->list_fk = ($request->input('data_type') == $value->element_fk) ? $request->input('list') : null;
+        $column->list_fk = $this->getListIdFromFormRequest($request);
         $column->data_type_fk = $request->input('data_type');
         $column->translation_fk = $request->input('translation');
         $column->description = $request->input('description');
@@ -207,9 +207,61 @@ class ColumnController extends Controller
      */
     public function destroy(Column $column)
     {
+        // Check for column mappings owning this column
+        if ($column->column_mapping->count()) {
+            return Redirect::to('admin/column')
+                ->with('warning', __('columns.still_owned_by'));
+        }
+        
         $column->delete();
         
         return Redirect::to('admin/column')
             ->with('success', __('columns.deleted'));
+    }
+
+    /**
+     * Get resource for AJAX autocompletion search field.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function autocomplete(Request $request)
+    {
+        $results = Column::select('column_id', 'description')
+            ->where('description', 'ILIKE', "%{$request->search}%")
+            ->orderBy('description')
+            ->limit(config('ui.autocomplete_results', 5))
+            ->get();
+        
+        $response = array();
+        foreach ($results as $result) {
+            $response[] = array(
+                "value" => $result->column_id,
+                "label" => $result->description,
+                "edit_url" => route('column.edit', $result->column_id),
+            );
+        }
+        
+        return response()->json($response);
+    }
+
+    /**
+     * Get the ID of the selected list if appropriate data type was chosen, null otherwise.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    private function getListIdFromFormRequest(Request $request)
+    {
+        // Get the element_id of data_type '_list' for validating the user's input
+        // TODO: check for list '_data_type' and maybe attribute
+        
+        if ($request->input('data_type') == Value::where('value', '_list_')->first()->element_fk) {
+            return $request->input('list');
+        }
+        if ($request->input('data_type') == Value::where('value', '_multi_list_')->first()->element_fk) {
+            return $request->input('list');
+        }
+        return null;
     }
 }
